+++ changelog.md
@@ -1,6 +1,7 @@
 ## SvCoreLib - Changelog
 ### Latest version: [1.14.2](#1142)
 
+
 <br><br>
 
 <details><summary><b>Version History - Click to expand</b></summary>