+++ changelog.md
@@ -50,6 +50,7 @@
     - `SelectionMenu` now supports EventEmitter's `.on("submit")` method
 - Fixes
     - Updated dependencies
+    - `seededRNG.validateSeed()` now returns false when a seed starts with `0` ([issue #34](https://github.com/Sv443-Network/SvCoreLib/issues/34))
     - Fixed missing argument in `system.inDebugger()` ([issue #37](https://github.com/Sv443-Network/SvCoreLib/issues/37))
     - `reserialize()` now keeps the type of the passed object ([issue #38](https://github.com/Sv443-Network/SvCoreLib/issues/38))